intptr_t is better to store a pointer (thanks Sam)
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@14510 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/protocols/proto_mpi/include/mpi_types.h
@@ -42,7 +42,7 @@
 #define MPI_THREAD_SERIALIZED 2
 #define MPI_THREAD_MULTIPLE 3
 
-typedef unsigned long MPI_Aint;
+typedef intptr_t MPI_Aint;
 
 typedef struct {
   int count;
